requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy>=1.20
pandas>=1.3
scipy>=1.7
scikit-learn>=1.0
scikit-image>=0.18
matplotlib>=3.4
tqdm>=4.60
Pillow>=8.0
imageio>=2.9
haven-ai>=0.7
